Added correct Folder Structure
Created Individual Components and fixed folder structure
