+++ App.js
@@ -1,25 +1,15 @@
 import React , {useState} from 'react';
-import { StyleSheet, Text, View , Button,TextInput} from 'react-native';
+import Login from './components/Login.js';
 
 export default function App() {
-  const [value , changeValue]  = useState ('Loggedout');
-  const [btnvalue , btnchangeValue]  = useState ('Login');
-
-  const changeText = ()=>{changeValue(value == 'Loggedout' ? 'LoggedIn' : 'Loggedout');
-  btnchangeValue(btnvalue == 'Login' ? 'Logout' : 'Login') }
+ 
   return (
-    <View style={styles.container}>
-      <Text>{value}</Text>
-      <TextInput
-          style={{height: 40}}
-          placeholder="Username"
-        />
-         <TextInput
-          style={{height: 40}}
-          placeholder="passowrd"
-        />
-      <Button title={btnvalue} onPress={changeText}/>
+
+     <View style={styles.container}>
+  
+   <Login/>
     </View>
+    
   );
 }
