+++ components/Category.js
@@ -1,0 +1,27 @@
+import React, {Component, useState} from 'react';
+import {StyleSheet, Text, Button, TextInput, View} from 'react-native';
+
+const Category=() => {
+
+  
+
+  
+        return(
+    <View style={styles.container}>
+      
+    </View>
+
+        );
+       
+    }
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  }
+});
+
+export default Category;
